+++ helpful.el
@@ -1317,8 +1317,9 @@ buffer."
         (buf nil)
         (pos nil)
         (opened nil)
-        ;; Skip running find-file-hook since it may prompt the user.
+        ;; Skip running hooks that may prompt the user.
         (find-file-hook nil)
+        (after-change-major-mode-hook nil)
         ;; If we end up opening a buffer, don't bother with file
         ;; variables. It prompts the user, and we discard the buffer
         ;; afterwards anyway.