+++ CHANGELOG.md
@@ -16,6 +16,8 @@ String values are now shown as literals by default.
 Fixed an issue with modes that override after-change-major-mode-hook,
 such as `global-hi-lock-mode`.
 
+Added bookmark support.
+
 # v0.16
 
 Improved wording when looking at aliases.